### 2020-03-06 #2142 @arfon
:wave: @richardjgowers - I think this also be an excellent paper for you to take on. What do you think?

/ cc @jedbrown
