### 2020-03-24 22212df
calculates distances using MDA.distance_array - ISSUE #30
