+++ CHANGELOG.rst
@@ -10,6 +10,11 @@ Upon version 0.8, and before version 1, SV2 major version increments are reflect
 Changelog
 =========
 
+0.8.10 (2020-xx-xx)
+-------------------
+
+* cli ``dist`` now uses ``MDA.distance_array`` function. ISSUE #30
+
 0.8.9 (2020-03-03)
 ------------------
 
