### 2020-04-01 #2142 @joaomcteixeira
Dear editor and reviewers,

Is there any news regarding restarting operations at JOSS after the suspension announced by @arfon referring to COVID-19? These are difficult times and I most deeply hope all you are in safe conditions.

A simple question: in parallel to this submission, I have continued developing the software. Should I update any data on the manuscript if a new PR is merged?

Thanks in advance,
