### 2020-04-01 #2142 @arfon
> A simple question: in parallel to this submission, I have continued developing the software. Should I update any data on the manuscript if a new PR is merged?

Given that the software has yet to go out for review I think it's fine to make small changes to the submission paper. Once the reviewers are working on the review I think it would be good to keep things stable.
