+++ setup.py
@@ -61,7 +61,7 @@ def _read(*names, **kwargs):
 
 setup(
     name='taurenmd',
-    version='0.8.10',
+    version='0.8.11',
     license='GNU GPLv2',
     description='A command-line interface for analysis routines in Molecular Dynamics data.',
     long_description='%s\n%s' % (
