### 2020-04-13 #2175 @whedon
[REVIEW]: taurenmd: A command-line interface for analysis of Molecular Dynamics simulations.
**Submitting author:** @joaomcteixeira (<a href="http://orcid.org/0000-0002-9113-0622">João M.C. Teixeira</a>)
**Repository:** <a href="https://github.com/joaomcteixeira/taurenmd" target ="_blank">https://github.com/joaomcteixeira/taurenmd</a>
**Version:** v0.8.9
