### 2020-04-13 #2175 @arfon
@amritagos, @Luthaf -
please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Any questions/concerns please let @richardjgowers know.
