### 2020-04-30 #2175 @amritagos
@richardjgowers @joaomcteixeira I am finished with my review.

`taurenmd` ties together some popular MD analysis libraries with an easy-to-use command-line interface. It is also possible to add new functionality, using the existing code structure. In particular, supported file formats include GROMACS trajectory files, which (to the best of my knowledge) are not supported by (OVITO)[https://www.ovito.org]. This is a neat, fairly well-documented software, which can make MD analysis accessible to newcomers to the field as well. I'm happy to recommend this for publication.
