### 2020-04-30 #2175 @joaomcteixeira
Dear @amritagos
I appreciate your comments of approval and positive feedback. I look forward to future interactions to continue developing `taurenmd` as a community-based software.
Cheers,
