### 2020-05-04 joaomcteixeira/taurenmd#35 @Luthaf
[BUG] `pip3 install taurenmd[all]` fails with zsh
I am using zsh as my main shell, not bash. I think it tries to interpret `[` and `]` as some sort of regex, meaning I get the following error

```
$ pip3 install taurenmd[all]
zsh: no matches found: taurenmd[all]
```

A simple fix is `pip3 install "taurenmd[all]"` (with quotes), which should also work with bash and other shells, so you may want to update the corresponding documentation.
