### 2020-05-04 joaomcteixeira/taurenmd#36 @Luthaf
[BUG] Broken installation with Anaconda
**Describe the bug**

Trying to install taurenmd using the Anaconda method outlined in the documentation, I get the following error:

```
$ taurenmd
Traceback (most recent call last):
  File "/usr/local/bin/taurenmd", line 5, in <module>
    from taurenmd.cli import maincli
  File "/usr/local/lib/python3.7/site-packages/taurenmd/cli.py", line 85, in <module>
    import taurenmd.cli_distances as cli_dist
  File "/usr/local/lib/python3.7/site-packages/taurenmd/cli_distances.py", line 48, in <module>
    from taurenmd.libs import libcli, libio, libmda, libplot  # noqa: F401
  File "/usr/local/lib/python3.7/site-packages/taurenmd/libs/libmda.py", line 14, in <module>
    import MDAnalysis as mda
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/__init__.py", line 194, in <module>
    from .lib import log
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/lib/__init__.py", line 40, in <module>
    from . import NeighborSearch
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/lib/NeighborSearch.py", line 37, in <module>
    from MDAnalysis.core.groups import AtomGroup, Atom
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/core/__init__.py", line 58, in <module>
    from . import AtomGroup
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/core/AtomGroup.py", line 28, in <module>
    from . import universe
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/core/universe.py", line 101, in <module>
    from .topology import Topology
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/core/topology.py", line 64, in <module>
    from .topologyattrs import Atomindices, Resindices, Segindices
  File "/Users/guillaume/code/mdanalysis/package/MDAnalysis/core/topologyattrs.py", line 38, in <module>
    import Bio.Seq
ModuleNotFoundError: No module named 'Bio'
```

**To Reproduce**

```bash
$ curl https://raw.githubusercontent.com/joaomcteixeira/taurenmd/master/requirements.yml -o taurenmdenv.yml
$ conda env create -f taurenmdenv.yml
$ conda activate taurenmd
# this was an error
$ pip install taurenmd
zsh: command not found: pip
# so I used this one
$ pip3 install taurenmd
$ taurenmd
# the above error
```

**Additional context**

I am running miniconda instead of a full anaconda distribution, not sure if that changes anything.

---

My guess is that there is a missing dependency in the PyPI package, or a missing conditional import.
