### 2020-05-04 joaomcteixeira/taurenmd#36 @Luthaf
Also, why not create an actual conda package with conda-forge? Using pip to install inside a conda environment feels strange to me.
