### 2020-05-04 170e1a9
answers issue 35
* updates requirements file, now install taurenmd directly from anaconda env
* updates documentation accordingly
* adds suggestions from issues 34 and 35.
