### 2020-05-04 joaomcteixeira/taurenmd#35 @joaomcteixeira
Hi @Luthaf
Thanks for your comments. I have added your explanation to the documentation, is to be merged in #37
