### 2020-05-04 joaomcteixeira/taurenmd#36 @joaomcteixeira
Hi @Luthaf
Thanks for your comments.

The reason I did not create an Anaconda package was because I did not felt the code of `taurenmd` required it. It is rather plain Python and of very easy install via `pip`. I do use `pip` in Anaconda to install many packages that are small and not distributed in Conda.

I added the `taurenmd` installation directly on the `requirements.yml` file, so it will install directly when the Anaconda env is created. This makes the installation process more natural. This updated is PR in #37.

Regarding your problem with the installation dependencies, this lies, to my understanding, outside the `taurenmd` scope, and it seams a problem with the `MDAnalysis` dependencies. I never had face it before, some questions so we can track down the problem:

* which system are you using?
* can you install `MDAnalysis` independently and use it outside `taurenmd`?

Best,
