### 2020-05-04 e4ed397
updated CHANGELOG to 0.8.12
