### 2020-05-04 3bdc54b
Bump version: 0.8.11 → 0.8.12
