### 2020-05-05 joaomcteixeira/taurenmd#36 @Luthaf
> Regarding your problem with the installation dependencies, this lies, to my understanding, outside the taurenmd scope, and it seams a problem with the MDAnalysis dependencies. I never had face it before, some questions so we can track down the problem

My bad, I forgot I had MDAnalysis installed as a development version locally, removing it fixes the issue. Sorry about that!
