+++ src/taurenmd/cli_rotations.py
@@ -124,7 +124,7 @@ def main(
         ):
     """Execute main client logic."""
     log.info(T('starting'))
-    
+
     topology = Path(topology)
     trajectories = [Path(t) for t in trajectories]
 
@@ -132,7 +132,7 @@ def main(
 
     log.info(T('transformation'))
     fSlice = libio.frame_slice(start=start, stop=stop, step=step)
-    
+
     origin_selection = ' or '.join(plane_selection)
     pABC_atomG = u.select_atoms(origin_selection)
     ABC_selections = plane_selection
@@ -142,18 +142,26 @@ def main(
     pC_atomG = u.select_atoms(ABC_selections[2])
 
     u.trajectory[0]
-    
+
     # defining the center of reference
-    pABC_cog = pABC_atomG.center_of_geometry()
+    opABC_cog = pABC_atomG.center_of_geometry()
     log.info(T('Original Center of Geometry'))
     log.info(S('for frame: 0'))
     log.info(S('for selection: {}', origin_selection))
-    log.info(S('pABC_cog: {}', pABC_cog))
-    
+    log.info(S('pABC_cog: {}', opABC_cog))
+
     log.info(T('Transfering'))
     log.info(S('all coordinates of reference frame to the origin 0, 0, 0'))
-    pABC_atomG.positions = pABC_atomG.positions - pABC_cog
+    pABC_atomG.positions = pABC_atomG.positions - opABC_cog
     log.info(S('COG in origin: {}', pABC_atomG.center_of_geometry()))
+
+    # defining the new center of reference
+    pABC_cog = pABC_atomG.center_of_geometry()
+    log.info(T('Origin Center of Geometry'))
+    log.info(S('for frame: 0'))
+    log.info(S('for selection: {}', origin_selection))
+    log.info(S('pABC_cog: {}', pABC_cog))
+   
 
     log.info(T('defining the reference axes'))
     pA_cog = pA_atomG.center_of_geometry()
@@ -165,14 +173,14 @@ def main(
     log.info(S('pC: {}', pC_cog))
 
     log.info(T('defining the normal vector to reference plane'))
-    ref_plane_normal = libcalc.calc_plane_normal(pA_cog, pB_cog, pC_cog)
+    ref_plane_normal = libcalc.calc_plane_normal(pA_cog, pB_cog, pABC_cog)
     log.info(S('plane normal: {}', ref_plane_normal))
     log.info(S('done'))
 
     log.info(T('defining the cross product vector'))
     ref_plane_cross = np.cross(pA_cog, ref_plane_normal)
     log.info(S('np cross: {}', ref_plane_cross))
-    
+
     roll_angles = []
     pitch_angles = []
     yaw_angles = []
@@ -180,95 +188,121 @@ def main(
     for i, _ts in enumerate(u.trajectory[fSlice]):
         print(f'.. working for frame :{i}')
 
+        # displaces all ts positions to the centre 0,0,0
+        # where the reference frame already is
         pABC_cog_ts = pABC_atomG.center_of_geometry()
-        pABC_atomG.positions = pABC_atomG.positions - pABC_cog_ts
+        pABC_atomG.positions = \
+            pABC_atomG.positions - pABC_cog_ts + ref_plane_normal
+        pABC_cog_ts = pABC_atomG.center_of_geometry()
 
         pA_cog_ts = pA_atomG.center_of_geometry()
-        pB_cog_ts = pB_atomG.center_of_geometry()
-        pC_cog_ts = pC_atomG.center_of_geometry()
 
-        ts_plane_normal = libcalc.calc_plane_normal(
-            pA_cog_ts,
-            pB_cog_ts,
[…]
