### 2020-05-10 5279b15
add tests
Tests cover libmdt.load_traj receiving list of trajectories
