### 2020-05-10 d5205e7
implementing partial
