+++ src/taurenmd/libs/libcalc.py
@@ -43,12 +43,12 @@ def mda_rmsd(
         ):
     """
     Calculate RMSDs observed for a selection.
-    
+
     Uses `MDAnalysis RMSD <https://www.mdanalysis.org/docs/documentation_pages/analysis/rms.html?highlight=rmsd#MDAnalysis.analysis.rms.RMSD>`_.
 
     Example
     -------
-    
+
     * Calculate RMSDs observed for the whole system along the whole trajectory.
 
         >>> mda_rmsd(universe)
@@ -76,7 +76,7 @@ def mda_rmsd(
     ref_frames : int, optional
         The reference frame against which calculate the RMSDs.
         Defaults to ``0``.
-    
+
     Returns
     -------
     Numpy Array
@@ -92,12 +92,12 @@ def mda_rmsd(
         Any exceptions that could come from MDAnalysis RMSF computation.
     """  # noqa: E501
     log.info(T('Calculating RMSDs'))
-    
+
     frame_slice = libio.evaluate_to_slice(value=frame_slice)
 
     log.info(S('for selection: {}', selection))
     log.info(S('for trajectory slice: {}', frame_slice))
-    
+
     R = mdaRMSD(
         universe,
         universe,
@@ -106,11 +106,11 @@ def mda_rmsd(
         ref_frame=ref_frame,
         verbose=False,
         )
-    
+
     R.run(verbose=False)
-    
+
     # rmsds[:, ii] = R.rmsd[:, 2][self._fslicer]
-    
+
     return R.rmsd[frame_slice, 2]
 
 
@@ -132,7 +132,7 @@ def mda_rmsf(
     frame_slice : any, optional
         Any argument that :py:func:`taurenmd.libs.libio.evaluate_to_slice` can receive.
         Defaults to ``None``, considers all frames.
-    
+
     Returns
     -------
     Numpy Array
@@ -145,11 +145,11 @@ def mda_rmsf(
         Any exceptions that could come from MDAnalysis RMSF computation.
     """  # noqa: E501
     log.info(T('Calculating RMSFs'))
-   
+
     frame_slice = libio.evaluate_to_slice(value=frame_slice)
-     
+
     log.info(S('for trajectory slice: {}', frame_slice))
-   
+
     R = mdaRMSF(
         atom_group,
         start=frame_slice.start,
@@ -157,9 +157,9 @@ def mda_rmsf(
         step=frame_slice.step,
         verbose=False,
         )
-   
+
     R.run()
-   
+
     return R.rmsf
 
 
@@ -201,9 +201,9 @@ def calc_plane_eq(p1, p2, p3):
     -------
     tuple of length 4
         The four parameters (a, b, c, d) that defined the plane equation:
-        
+
         .. math::
-        
+
             ax + by + cz = d
     """  # noqa: E501
     cp = calc_plane_normal(p1, p2, p3)
@@ -222,7 +222,7 @@ def calc_planes_angle(a1, b1, c1, a2, b2, c2, aunit='radians'):
     .. math::
 
         a1*x + b1*y + c1*z + d = 0
-        
+
         a2*x + b2*y + c2*z + d = 0
 
     `Read further <from: https://www.geeksforgeeks.org/angle-between-two-planes-in-3d/>`_.
@@ -231,7 +231,7 @@ def calc_planes_angle(a1, b1, c1, a2, b2, c2, aunit='radians'):
     ----------
     a1, b1, c1, a2, b2, c2 : float
         Plane parameters
-    
+
     angle : str, optional
         ``degrees`` returns angle quantity in degrees, else returns
         in radians.
@@ -267,7 +267,7 @@ def generate_quaternion_rotations(
     Rotates a `vector` around an `axis` for a series of angles.
[…]
