+++ src/taurenmd/cli_rotations.py
@@ -154,10 +154,11 @@ def main(
     log.info(T('Transfering'))
     log.info(S('all coordinates of reference frame to the origin 0, 0, 0'))
     pABC_atomG.positions = pABC_atomG.positions - opABC_cog
-    log.info(S('COG in origin: {}', pABC_atomG.center_of_geometry()))
 
     # defining the new center of reference
-    pABC_cog = pABC_atomG.center_of_geometry()
+    pABC_cog = pABC_atomG.center_of_geometry().copy()
+    log.info(S('COG in origin: {}', pABC_cog))
+
     log.info(T('Origin Center of Geometry'))
     log.info(S('for frame: 0'))
     log.info(S('for selection: {}', origin_selection))
@@ -165,13 +166,13 @@ def main(
 
 
     log.info(T('defining the reference axes'))
-    pA_cog = pA_atomG.center_of_geometry()
-    pB_cog = pB_atomG.center_of_geometry()
-    pC_cog = pC_atomG.center_of_geometry()
+    pA_cog = pA_atomG.center_of_geometry().copy()
+    pB_cog = pB_atomG.center_of_geometry().copy()
+    pC_cog = pC_atomG.center_of_geometry().copy()
     log.info(S('plane points definition:'))
     log.info(S('pA: {}', pA_cog))
     log.info(S('pB: {}', pB_cog))
-    log.info(S('pC: {}', pC_cog))
+    #log.info(S('pC: {}', pC_cog))
 
     log.info(T('defining the normal vector to reference plane'))
     ref_plane_normal = libcalc.calc_plane_normal(pA_cog, pB_cog, pABC_cog)
@@ -182,129 +183,108 @@ def main(
     ref_plane_cross = np.cross(pA_cog, ref_plane_normal)
     log.info(S('np cross: {}', ref_plane_cross))
 
-    roll_angles = []
-    pitch_angles = []
-    yaw_angles = []
+    #roll_angles = []
+    #pitch_angles = []
+    #yaw_angles = []
+
+    total_frames = len(u.trajectory[fSlice])
+    sizet = (total_frames, 3)
+    roll_vectors = np.empty(sizet, dtype=np.float64)
+    pitch_vectors = np.empty(sizet, dtype=np.float64)
+    yaw_vectors = np.empty(sizet, dtype=np.float64)
 
     for i, _ts in enumerate(u.trajectory[fSlice]):
         print(f'.. working for frame :{i}')
 
-        # displaces all ts positions to the centre 0,0,0
-        # where the reference frame already is
-        pABC_cog_ts = pABC_atomG.center_of_geometry()
-        pABC_atomG.positions = \
-            pABC_atomG.positions - pABC_cog_ts + ref_plane_normal
-        pABC_cog_ts = pABC_atomG.center_of_geometry()
+        pABC_cog_ts = pABC_atomG.center_of_geometry().copy()
+        pABC_atomG.positions = pABC_atomG.positions - pABC_cog_ts
+        #pABC_cog_ts = pABC_atomG.center_of_geometry().copy()
 
-        pA_cog_ts = pA_atomG.center_of_geometry()
+        ts_positions = pABC_atomG.positions.copy()
 
-        p1 = pA_cog
-        p2 = pABC_cog
-        p3 = pABC_cog_ts
-        #p4 = pA_cog_ts
-        p4 = pA_cog
-        coords = np.array([p1, p2, p3, p4])
+        # get roll
+        pABC_atomG.positions = ts_positions + ref_plane_normal
+        roll_vectors[i, :] = pA_atomG.center_of_geometry().copy()
+        pABC_atomG.positions = ts_positions
 
-        torsion = libcalc.calc_torsion_angles(coords)
[…]
