### 2020-05-12 fb0ef22
corrects brackets
Commit direct to master. Travis isort configuration differs
in master from that of #39. To address in #38.
