### 2020-05-12 e65f07a
Updates tox.ini
* updates tox according to #38
* updates isort installation to follow latest on their gitrepository
* creates a new isort env in tox to apply isort changes sugggestions
