+++ src/taurenmd/cli_nosol.py
@@ -25,8 +25,9 @@ Remove all solvent except for NA atoms:
 import argparse
 import functools
 
-import taurenmd.core as tcore
-from taurenmd import _BANNER, log
+from taurenmd import _BANNER
+from taurenmd import core as tcore
+from taurenmd import log
 from taurenmd.libs import libcli, libio, libmdt
 from taurenmd.logger import S, T
 
