+++ src/taurenmd/cli_pangle.py
@@ -36,8 +36,9 @@ where ``[...]`` is the previous command example.
 import argparse
 import functools
 
-import taurenmd.core as tcore
-from taurenmd import _BANNER, Path, log
+from taurenmd import _BANNER, Path
+from taurenmd import core as tcore
+from taurenmd import log
 from taurenmd.libs import libcalc, libcli, libio, libmda, libplot
 from taurenmd.logger import S, T
 
