+++ tox.ini
@@ -13,7 +13,7 @@ ignore_basepython_conflict = true
 basepython =
     {py36}: {env:TOXPYTHON:python3.6}
     {py37,docs}: {env:TOXPYTHON:python3.7}
-    {check,radon,codecov,coveralls}: {env:TOXPYTHON:python3}
+    {check,radon,codecov,coveralls,isort}: {env:TOXPYTHON:python3}
 passenv = *
 
 [testenv:py36]
@@ -43,7 +43,7 @@ conda_channels =
 commands_pre =
     coverage erase
 commands =
-    {posargs:pytest --cov --cov-report=term-missing --cov-append --cov-config=.coveragerc -vv tests}
+    pytest --cov --cov-report=term-missing --cov-append --cov-config=.coveragerc -vv {posargs}
 commands_post = 
     coverage report
     coverage html
@@ -67,18 +67,24 @@ deps =
     flake8-bugbear
     readme-renderer
     pygments
-    isort
+    #isort
+    git+git://github.com/timothycrosley/isort.git
     bumpversion
 #usedevelop = false
 skip_install = true
 commands =
     python setup.py check --strict --metadata --restructuredtext
     check-manifest {toxinidir}
-    flake8 src tests setup.py docs
-    isort --verbose --check-only --diff --recursive src tests setup.py
+    flake8 {posargs:src tests setup.py docs}
+    isort --verbose --check-only --diff {posargs:src tests setup.py}
     bumpversion --dry-run --verbose --allow-dirty patch
     bumpversion --dry-run --verbose --allow-dirty minor
     bumpversion --dry-run --verbose --allow-dirty major
+
+[testenv:isort]
+skip_install = true
+deps = git+git://github.com/timothycrosley/isort.git
+commands = isort --verbose --diff {posargs:src tests setup.py}
 
 [testenv:radon]
 deps = radon
@@ -145,7 +151,7 @@ exclude = src/taurenmd/__init__.py
 skip = __init__.py
 line_length = 80
 indent = 4
-multi_line_output = 3
+multi_line_output = 8
 include_trailing_comma = true
 lines_after_imports = 2
 sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
@@ -181,4 +187,3 @@ python_files =
 #    --tb=short
 testpaths =
     tests
-
