+++ tox.ini
@@ -151,7 +151,7 @@ exclude = src/taurenmd/__init__.py
 skip = __init__.py
 line_length = 80
 indent = 4
-multi_line_output = 8
+multi_line_output = VERTICAL_HANGING_INDENT_BRACKET
 include_trailing_comma = true
 lines_after_imports = 2
 sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
