+++ CHANGELOG.rst
@@ -10,6 +10,11 @@ Upon version 0.8, and before version 1, SV2 major version increments are reflect
 Changelog
 =========
 
+0.8.14 (2020-05-12)
+-------------------
+
+* Updates `tox.ini` file for Continuous Integration (PR #40)
+
 0.8.13 (2020-05-12)
 -------------------
 
