+++ docs/conf.py
@@ -43,7 +43,7 @@ project = 'taurenmd'
 year = '2019'
 author = 'Joao MC Teixeira'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '0.8.13'
+version = release = '0.8.14'
 
 todo_include_todos = True
 pygments_style = 'trac'
