### 2020-05-14 joaomcteixeira/taurenmd#41 @Luthaf
[DOCUMENTATION] test errors while going through contribution guideline
I am trying to follow https://taurenmd.readthedocs.io/en/latest/contributing.html#code-development, and I got a few errors while running tox:

- `ERROR:  py36: InterpreterNotFound: python3.6`: would there be a way to automatically check if python 3.6 is installed instead of erroring if it can not be found?
- error related to missing dependencies, they should perhaps be added to `tox.ini`

```
___________________________________________________________________________________ ERROR collecting tests/test_clis.py ____________________________________________________________________________________
ImportError while importing test module '/Users/guillaume/code/tests/taurenmd/tests/test_clis.py'.
Traceback:
.tox/py37/lib/python3.7/site-packages/_pytest/python.py:511: in _importtestmodule
    mod = self.fspath.pyimport(ensuresyspath=importmode)
.tox/py37/lib/python3.7/site-packages/py/_path/local.py:701: in pyimport
    __import__(modname)
.tox/py37/lib/python3.7/site-packages/_pytest/assertion/rewrite.py:152: in exec_module
    exec(co, module.__dict__)
tests/test_clis.py:12: in <module>
    from taurenmd import __main__ as main_module
.tox/py37/lib/python3.7/site-packages/taurenmd/__main__.py:10: in <module>
    from taurenmd.cli import maincli
.tox/py37/lib/python3.7/site-packages/taurenmd/cli.py:86: in <module>
    from taurenmd import cli_distances as cli_dist
.tox/py37/lib/python3.7/site-packages/taurenmd/cli_distances.py:49: in <module>
    from taurenmd.libs import libcli, libio, libmda, libplot  # noqa: F401
.tox/py37/lib/python3.7/site-packages/taurenmd/libs/libmda.py:14: in <module>
    import MDAnalysis as mda
E   ModuleNotFoundError: No module named 'MDAnalysis'
__________________________________________________________________________________ ERROR collecting tests/test_libcalc.py __________________________________________________________________________________
ImportError while importing test module '/Users/guillaume/code/tests/taurenmd/tests/test_libcalc.py'.
Traceback:
.tox/py37/lib/python3.7/site-packages/_pytest/python.py:511: in _importtestmodule
    mod = self.fspath.pyimport(ensuresyspath=importmode)
.tox/py37/lib/python3.7/site-packages/py/_path/local.py:701: in pyimport
    __import__(modname)
.tox/py37/lib/python3.7/site-packages/_pytest/assertion/rewrite.py:152: in exec_module
    exec(co, module.__dict__)
tests/test_libcalc.py:2: in <module>
    import MDAnalysis as mda
E   ModuleNotFoundError: No module named 'MDAnalysis'
__________________________________________________________________________________ ERROR collecting tests/test_libmda.py ___________________________________________________________________________________
ImportError while importing test module '/Users/guillaume/code/tests/taurenmd/tests/test_libmda.py'.
Traceback:
.tox/py37/lib/python3.7/site-packages/_pytest/python.py:511: in _importtestmodule
    mod = self.fspath.pyimport(ensuresyspath=importmode)
.tox/py37/lib/python3.7/site-packages/py/_path/local.py:701: in pyimport
    __import__(modname)
.tox/py37/lib/python3.7/site-packages/_pytest/assertion/rewrite.py:152: in exec_module
    exec(co, module.__dict__)
tests/test_libmda.py:4: in <module>
    import MDAnalysis as mda
E   ModuleNotFoundError: No module named 'MDAnalysis'
__________________________________________________________________________________ ERROR collecting tests/test_libmdt.py ___________________________________________________________________________________
ImportError while importing test module '/Users/guillaume/code/tests/taurenmd/tests/test_libmdt.py'.
Traceback:
.tox/py37/lib/python3.7/site-packages/_pytest/python.py:511: in _importtestmodule
    mod = self.fspath.pyimport(ensuresyspath=importmode)
.tox/py37/lib/python3.7/site-packages/py/_path/local.py:701: in pyimport
    __import__(modname)
.tox/py37/lib/python3.7/site-packages/_pytest/assertion/rewrite.py:152: in exec_module
    exec(co, module.__dict__)
tests/test_libmdt.py:2: in <module>
    import mdtraj as md
E   ModuleNotFoundError: No module named 'mdtraj'
```
