### 2020-05-14 30b0f38
revisiting algorithm
