### 2020-05-14 joaomcteixeira/taurenmd#41 @joaomcteixeira
Hi @Luthaf

* regarding the missing interpreter, could you please add the following line to the `tox.ini` and see if it maintains the same error?
  * `skip_missing_interpreters=true`, add it just after [here](https://github.com/joaomcteixeira/taurenmd/blob/acff0c01f02d9202bad312695796d42d4ae54d99/tox.ini#L10)
  * I have created a new conda environment to attempt to reproduce your error but I could not. `tox` always finds the interpreters.
  * Also notice that, despite the *missing interpreter* errors that are known to happen with `tox`, all the environments are tested later in `Travis-CI`.

* regarding the dependencies, `tox.ini` already has information to download all dependencies, and `tox-conda` should manage all that. Have you installed `tox-conda` as well?
  * also, please notice that development of `taurenmd` is bound to Anaconda, at least for now:

> ([from the documentation](https://taurenmd.readthedocs.io/en/latest/contributing.html#general-instructions))Though taurenmd can be used without relying on Anaconda to manage its dependencies, there are actually some dependencies that are only available in the Anaconda ecosystem. Therefore, the full development of taurenmd is bound to Anaconda until this situation is solved, you can read further on this on our installation page.

Please notice also that `taurenmd` current version is [`0.8.14`](https://github.com/joaomcteixeira/taurenmd/releases/tag/v0.8.14), and the one proposed in the article is [`0.8.9`](https://github.com/joaomcteixeira/taurenmd/releases/tag/v0.8.9), however, there have **not** been crucial changed to the `tox.ini` file during these updates. From the `master` branch you can see the changes in the `tox.ini` file since `0.8.9` with the following command:

```
git diff 40cf753dfb3afe0c3fcbc49696d0982314fad8f2 acff0c01f02d9202bad312695796d42d4ae54d99 tox.ini
```

Should the documentation be clarified in this regard?
