+++ tests/test_libcalc.py
@@ -42,13 +42,13 @@ def test_mda_rmsf():
             np.array([0.0, 0.0, 0.0]),
             np.array([1.0, 0.0, 0.0]),
             np.array([0.0, 1.0, 0.0]),
-            np.array([0.0, 0.0, -1.0])
+            np.array([0.0, 0.0, 1.0])
             ),
         (
             np.array([0.0, 0.0, 0.0]),
             np.array([0.0, -1.0, 0.0]),
             np.array([0.0, 0.0, 1.0]),
-            np.array([1.0, 0.0, 0.0])
+            np.array([-1.0, 0.0, 0.0])
             ),
         ],
     )
@@ -65,13 +65,13 @@ def test_calc_plane_normal(p1, p2, p3, expected):
             np.array([0.0, 0.0, 0.0]),
             np.array([1.0, 0.0, 0.0]),
             np.array([0.0, 1.0, 0.0]),
-            (0.0, 0.0, -1.0, 0.0),
+            (0.0, 0.0, 1.0, 0.0),
             ),
         (
             np.array([0.0, 0.0, 0.0]),
             np.array([0.0, -1.0, 0.0]),
             np.array([0.0, 0.0, 1.0]),
-            (1.0, 0.0, 0.0, 0.0)
+            (-1.0, 0.0, 0.0, 0.0)
             ),
         ],
     )
