+++ src/taurenmd/cli_rotations.py
@@ -69,14 +69,13 @@ In the case of an homotrimer, define the axis and the origin on the trimers:
 """  # noqa: E501
 import argparse
 import functools
-import pprint
 
 import numpy as np
 
 from taurenmd import _BANNER, Path
 from taurenmd import core as tcore
 from taurenmd import log
-from taurenmd.libs import libcalc, libcli, libio, libmda  # noqa: F401
+from taurenmd.libs import libcalc, libcli, libio, libmda, libplot  # noqa: F401
 from taurenmd.logger import S, T
 
 
@@ -105,8 +104,10 @@ libcli.add_topology_arg(ap)
 libcli.add_trajectories_arg(ap)
 libcli.add_plane_selection_arg(ap)
 libcli.add_angle_unit_arg(ap)
+libcli.add_reference_frame_arg(ap)
 libcli.add_slice_arg(ap)
 libcli.add_data_export_arg(ap)
+libcli.add_plot_arg(ap)
 
 
 def _ap():
@@ -118,10 +119,13 @@ def main(
         trajectories,
         plane_selection,
         aunit='degrees',
+        ref_frame=0,
         start=None,
         stop=None,
         step=None,
         export=False,
+        plot=False,
+        plotvars=None,
         **kwargs,
         ):
     """Execute main client logic."""
@@ -129,40 +133,38 @@ def main(
 
     topology = Path(topology)
     trajectories = map(Path, trajectories)
-
     u = libmda.load_universe(topology, *trajectories)
 
-    log.info(T('transformation'))
+    log.info(T('slicing'))
     fSlice = libio.frame_slice(start=start, stop=stop, step=step)
 
-    origin_selection = ' or '.join(plane_selection)
-    pABC_atomG = u.select_atoms(origin_selection)
+    log.info(T('computing planes'))
+    log.info(S('for reference frame: {}', ref_frame))
+
     ABC_selections = plane_selection  # variable renaming
+    origin_selection = ' or '.join(ABC_selections)
+    log.info(S('for selection: {!r}', origin_selection))
 
     # p stands for point
     # atomG stands for MDAnalysis AtomGroup
+    pABC_atomG = u.select_atoms(origin_selection)
     pA_atomG = u.select_atoms(ABC_selections[0])
     pB_atomG = u.select_atoms(ABC_selections[1])
     pC_atomG = u.select_atoms(ABC_selections[2])
 
-    u.trajectory[0]
+    u.trajectory[ref_frame]
 
     # defining the center of reference
     opABC_cog = pABC_atomG.center_of_geometry()
-    log.info(T('Original Center of Geometry'))
-    log.info(S('for frame: 0'))
-    log.info(S('for selection: {!r}', origin_selection))
-    log.info(S('pABC center of geometry: {}', opABC_cog))
+    log.info(S('Original pABC center of geometry: {}', opABC_cog))
 
-    log.info(T('Transfering'))
-    log.info(S('all coordinates of reference frame to the origin 0, 0, 0'))
+    log.info(S(
+        'transferring all coordinates of '
+        'reference frame to the origin (0, 0, 0)'
+        ))
     pABC_atomG.positions = pABC_atomG.positions - opABC_cog
-
-    # defining the new center of reference
     pABC_cog = pABC_atomG.center_of_geometry().copy()
-    log.info(T('Origin Center of Geometry'))
-    log.info(S('for frame: 0'))
-    log.info(S('for selection: {}', origin_selection))
+    log.info(S('New origin Center of Geometry'))
     log.info(S('pABC_cog: {}', pABC_cog))
 
     log.info(T('defining the reference axes'))
@@ -174,27 +176,24 @@ def main(
[…]
