### 2020-05-15 25fcea5
Adds input sort argument to all CLIs
* CLI `imagemol` conflicts with `-i` option for `protocol`, therefore,
`-i` for `--protocol` was removed. This should be accompained by a
`major` version change.
