### 2020-05-15 d67c907
adds -i to all arg_parser in CLIs
