+++ src/taurenmd/core.py
@@ -46,10 +46,10 @@ class Path(type(_Path())):
 ref_taurenmd = "* Cite taurenmd according to: https://taurenmd.readthedocs.io/en/latest/citing.html\n"  # noqa: E501
 """How to cite taurenmd project."""
 
-ref_mdt = "* MD data is accessed and/or processed using [MDTraj](https://mdtraj.org/)\n."  # noqa: E501
+ref_mdt = "* MD data was accessed and/or processed using [MDTraj](https://mdtraj.org/)\n."  # noqa: E501
 """Command-line docstring to reference MDTraj package."""
 
-ref_mda = "* MD data is accessed using [MDAnalysis](https://www.mdanalysis.org).\n"  # noqa: E501
+ref_mda = "* MD data was accessed using [MDAnalysis](https://www.mdanalysis.org).\n"  # noqa: E501
 """Command-line docstring to reference MDAnalysis package."""
 
 ref_mda_selection = "* selection commands follow MDAnalysis [selection nomenclature](https://www.mdanalysis.org/docs/documentation_pages/selections.html#).\n"  # noqa: E501
@@ -61,13 +61,13 @@ ref_mda_unwrap = "* unwrap performed by MDAnalysis [unwrap](https://www.mdanalys
 ref_mda_alignto = "* align performed by MDAnalysis [unwrap](https://www.mdanalysis.org/docs/documentation_pages/analysis/align.html?highlight=alignto#MDAnalysis.analysis.align.alignto).\n"  # noqa: E501
 """Command-line docstring to reference MDAnalysis alignto function."""
 
-ref_plottemplates_param = "* plotting is performed by [python-bioplottemplates plot param function](https://python-bioplottemplates.readthedocs.io/en/latest/reference/plots.html#bioplottemplates.plots.param.plot).\n"  # noqa: E501
+ref_plottemplates_param = "* plotting was performed by [python-bioplottemplates plot param function](https://python-bioplottemplates.readthedocs.io/en/latest/reference/plots.html#bioplottemplates.plots.param.plot).\n"  # noqa: E501
 """Command-line docstring to reference python-bioplottemplates param plot."""
 
-ref_plottemplates_labeldots = "* plotting is performed by [python-bioplottemplates plot labeldots function](https://python-bioplottemplates.readthedocs.io/en/latest/reference/plots.html#bioplottemplates.plots.label_dots.plot).\n"  # noqa: E501
+ref_plottemplates_labeldots = "* plotting was performed by [python-bioplottemplates plot labeldots function](https://python-bioplottemplates.readthedocs.io/en/latest/reference/plots.html#bioplottemplates.plots.label_dots.plot).\n"  # noqa: E501
 """Command-line docstring to reference python-bioplottemplates labeldots plot."""  # noqa: E501
 
-ref_pyquaternion = "* Quaterion operations are performed with [pyquaterion](http://kieranwynn.github.io/pyquaternion/).\n"  # noqa: E501
+ref_pyquaternion = "* Quaterion operations were performed with [pyquaterion](http://kieranwynn.github.io/pyquaternion/).\n"  # noqa: E501
 """Command-line docstring to reference pyquaterion lib."""
 
 ref_numpy = '* Matrix operations were performed with [Numpy](https://www.scipy.org/citing.html).'  # noqa: E501
