+++ src/taurenmd/libs/libmdt.py
@@ -79,19 +79,41 @@ def load_traj(topology, trajectories, insort=False):
         trajs = os.fspath(trajectories)
 
     libio.report_input(topology, trajs)
-
-    topp = Path(topology)
-    if topp.suffix == '.cif' and SIMTK:
-        mol = app.PDBxFile(topp.str())
-        top = mdtraj.Topology.from_openmm(mol.topology)
-    elif topp.suffix == '.cif' and not SIMTK:
-        _log_simtkimport_error()
-    else:
-        top = topp.str()
-
+    top = attempt_to_load_top_from_simtk(topology)
     mdtrajectory = mdtraj.load(trajs, top=top)
 
     return mdtrajectory
+
+
+@libcli.add_reference(tcore.ref_openmm)
+def attempt_to_load_top_from_simtk(topology):
+    """
+    Load topology from SIMTK.
+
+    Parameters
+    ----------
+    topology : str or Path
+
+    Returns
+    -------
+    topology from mdtraj.Topology.from_openmm`
+
+    Raises
+    ------
+    Dependency error from :func:`_log_simtkimport_error`, program
+    halts.
+    """
+    topp = Path(topology)
+
+    if topp.suffix == '.cif' and SIMTK:
+        mol = app.PDBxFile(topp.str())
+        return mdtraj.Topology.from_openmm(mol.topology)
+
+    elif topp.suffix == '.cif' and not SIMTK:
+        _log_simtkimport_error()
+
+    else:
+        return topp.str()
 
 
 @libcli.add_reference(tcore.ref_mdt)
