+++ tests/test_libmdt.py
@@ -35,6 +35,16 @@ def test_load_traj_cif():
     assert len(traj) == 10
 
 
+def test_attempt_load_cif_SIMTK_1():
+    """Test attempt load CIF."""
+    libmdt.attempt_to_load_top_from_simtk(toptest_cif)
+
+
+def test_attempt_load_cif_SIMTK_2():
+    """Test attempt load CIF."""
+    assert toptest.str() == libmdt.attempt_to_load_top_from_simtk(toptest)
+
+
 def test_load_traj_cif_import_error():
     """Test loading traj."""
     libmdt.SIMTK = False
