+++ CHANGELOG.rst
@@ -10,6 +10,11 @@ Upon version 0.8, and before version 1, SV2 major version increments are reflect
 Changelog
 =========
 
+0.9.1 (2020-05-15)
+------------------
+
+* Improves log in `.taurenmd.cmd` (PR #43)
+
 0.9.0 (2020-05-15)
 ------------------
 
