### 2020-05-15 271992c
added function to trajedit
