### 2020-05-17 7cf89d7
adds bars to clis
