+++ src/taurenmd/cli_pangle.py
@@ -96,7 +96,7 @@ def main(
         ):
     """Execute main client logic."""
     log.info(T('calculating angles'))
-    
+
     topology = Path(topology)
     trajectories = [Path(t) for t in trajectories]
 
@@ -122,26 +122,31 @@ def main(
         reference_point_3,
         )
     log.info(S('the equation is {}x + {}y + {}z = {}', ra, rb, rc, rd))
-    
+
     log.info(T('Calculating angles'))
     angles = []
-    for _ts in u.trajectory[frame_slice]:
 
-        point1 = u.select_atoms(plane_selection[0]).center_of_geometry()
-        point2 = u.select_atoms(plane_selection[1]).center_of_geometry()
-        point3 = u.select_atoms(plane_selection[2]).center_of_geometry()
-        
-        a, b, c, d = libcalc.calc_plane_eq(point1, point2, point3)
+    trajlen = len(u.trajectory[frame_slice])
+    with libcli.ProgressBar(trajlen, suffix='frames') as pb:
+        for _ts in u.trajectory[frame_slice]:
 
-        angles.append(
-            libcalc.calc_planes_angle(
-                ra, rb, rc, a, b, c,
-                aunit=aunit,
+            point1 = u.select_atoms(plane_selection[0]).center_of_geometry()
+            point2 = u.select_atoms(plane_selection[1]).center_of_geometry()
+            point3 = u.select_atoms(plane_selection[2]).center_of_geometry()
+
+            a, b, c, d = libcalc.calc_plane_eq(point1, point2, point3)
+
+            angles.append(
+                libcalc.calc_planes_angle(
+                    ra, rb, rc, a, b, c,
+                    aunit=aunit,
+                    )
                 )
-            )
-    
+
+            pb.increment()
+
     log.info(S('calculated a total of {} angles.', len(angles)))
-   
+
     if export:
         libio.export_data_to_file(
             list(range(len(u.trajectory))[frame_slice]),
