### 2020-05-17 7c82037
test 100%
* add indexerror test
* trick line to allow 100% coverage
