### 2020-05-22 joaomcteixeira/taurenmd#41 @Luthaf
> skip_missing_interpreters=true, add it just after here

This fixed the issue indeed.

> Have you installed tox-conda as well?

I was missing tox-conda, with it everything seems to work fine. I had seen the documentation regarding Anaconda, and I did not understand it as "you can **only** run tests and develop the code with Anaconda". If pip/PyPi is not supported, I think that's fine but should be made clearer.

But I though that only OpenMM was not available on conda, right? What prevents you from changing tox.ini from

```
deps =
    pytest
    pytest-travis-fold
    pytest-cov
    coverage
    bioplottemplates
    pyquaternion
conda_deps =
    mdanalysis
    mdtraj
    openmm
    matplotlib
```

to

```
deps =
    pytest
    pytest-travis-fold
    pytest-cov
    coverage
    bioplottemplates
    pyquaternion
    mdanalysis
    mdtraj
    matplotlib
conda_deps =
    openmm
```
