### 2020-05-22 joaomcteixeira/taurenmd#45 @Luthaf
JOSS-review: comments about the paper
> taurenmd is deployed in the Python ecosystem and is available for direct download at PyPI
> (https://pypi.org/project/taurenmd/):
>                       $ pip3 install taurenmd[all]

Given the issues with Anaconda/PyPi dependencies, I would refrain from giving the installation command directly in the paper, and rather point users to the documentation for this.

I also think that taurenmd would benefit from being also distributed as a conda package as well as a pip one. I personally prefer not to mix conda-managed packages with pip ones; as you can realise from the issues I've been opening =) Since Python package management is such a mess, I prefer to keep two separated package trees to keep it under control.

Whenever I see a package telling me to install it with pip, I assume it works fully on PyPi, with no dependencies on conda universe; and inversely. I am sure some future users of taurenmd will make the same distinction.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I did not find such comparison in the paper, in particular to the `gmx` family of tools distributed with GROMACS (I guess taurenmd would be easier to extend but slower/available on less platforms); pytraj (https://github.com/Amber-MD/pytraj); or graphical interface tools such as the analysis plugins of VMD, PyMol, Chimera.

Since taurenmd seems to be targeted at non-programmers, I think a justification of its use case compared to GUI tools is needed, since a command line tool is less beginner-friendly than a GUI. In particular, MDAnalysis seems to be usable [within PyMol directly](https://nms.kcl.ac.uk/lorenz.lab/wp/?p=1768).
