### 2020-05-23 #2175 @joaomcteixeira
Hi @richardjgowers ,
@Luthaf made some comments on the review as issues in `taurenmd` repository, I am now addressing them: https://github.com/joaomcteixeira/taurenmd/issues/35, https://github.com/joaomcteixeira/taurenmd/issues/36, https://github.com/joaomcteixeira/taurenmd/issues/41, https://github.com/joaomcteixeira/taurenmd/issues/45
