### 2020-05-23 joaomcteixeira/taurenmd#35 @joaomcteixeira
refers to https://github.com/openjournals/joss-reviews/issues/2175
