### 2020-05-23 8150f46
Clarifies CONTRIBUTION instructions
* Clarifies text in CONTRIBUTION.rst
* Creates new `requirements-dev.yml`
* Addresses #41

* bumpversion: `patch`
