+++ CONTRIBUTING.rst
@@ -41,58 +41,52 @@ General instructions
 --------------------
 
 Though *taurenmd* can be used without relying on Anaconda to manage its dependencies,
-there are actually some dependencies that are only available in the Anaconda ecosystem.
-Therefore, the full development of *taurenmd* is bound to Anaconda until this situation is solved,
-you can read further on this in our :ref:`installation page <Installation>`.
+some dependencies are only available in the Anaconda ecosystem. Therefore,
+although *taurenmd* can be used almost entirely without relying on Anaconda,
+its development is bound, currently, to Anaconda.
+You can read further on this on our installation page.
 
-Also, by using Conda, our CI integration is much faster because the heavy dependencies
-are handled just faster in Anaconda than in pure PyPI.
+Also, by using Conda, our CI integration is much faster because the heavy
+dependencies are handled just faster in Anaconda than in pure PyPI. However,
+it is possible to do some gymnastics and avoid using Conda and
+still contribute to the development *taurenmd*; if this is your case,
+please write us an issue, and we will help you out.
 
-For these reasons, the development of *taurenmd* is bound to Anaconda. However, it is
-possible to make some gymnastics and avoid using Conda and still contribute
-to the develop *taurenmd*; if this is your case, please write us an issue and
-we will help you out.
+From this moment on, we will assume you are using Anaconda as your Python package manager.
+Also, we are assuming you have `Git <https://git-scm.com/>`_ installed, and you are running a Linux machine
+– we are not aware of the possible divergences of these instructions in other OSes.
 
-From this moment on we will assume you are using Anaconda as your Python package manager.
+1. Create a new environment with the *taurenmd* dependencies, **only its dependencies**, running the following 3 commands::
 
-1. Create a new environment and install *taurenmd* dependencies, **only its dependencies**, following the instructions
-for :ref:`Anaconda in our Installation <With Anaconda>` page. **Important**, change the name of the
-environment in the ``yml`` file before actually creating the environment, for example,
-name it ``taurenmddev``.
+    curl https://raw.githubusercontent.com/joaomcteixeira/taurenmd/master/requirements-dev.yml -o taurenmddev.yml
+    conda env create -f taurenmdenv.yml
+    # Activate the conda taurenmddev environment
+    conda activate taurenmddev
 
-    1.1. Remember to activate the new environment before proceeding to the installation.
+2. Fork `taurenmd <https://github.com/joaomcteixeira/taurenmd>`_ (look for the "Fork" button on the top right corner of `our repository <https://github.com/joaomcteixeira/taurenmd>`_).
 
-2. *taurenmd* relies on `tox <https://tox.readthedocs.io/en/latest/>`_ to manage continuous integration (CI) and collaborative development; install it together with `tox-conda <https://github.com/tox-dev/tox-conda>`_::
+3. `Clone <https://help.github.com/en/github/creating-cloning-and-archiving-repositories/cloning-a-repository>`_ your forked repository to your local machine::
 
-    conda install -c conda-forge tox
-    conda install -c conda-forge tox-conda
+    git clone https://github.com/YOUR-USER-NAME/taurenmd.git <destination folder>
 
-3. Fork `taurenmd <https://github.com/joaomcteixeira/taurenmd>`_ (look for the "Fork" button on the top right corner of `our repository <https://github.com/joaomcteixeira/taurenmd>`_).
-
-
-4. `Clone <https://help.github.com/en/github/creating-cloning-and-archiving-repositories/cloning-a-repository>`_ your forked repository to your local machine::
-
-    git clone https://github.com/YOUR-USER-NAME/taurenmd.git <destination folder> 
-
-5. Navigate to the fork folder and create a branch for local development::
[…]
