### 2020-05-23 joaomcteixeira/taurenmd#41 @joaomcteixeira
Hi @Luthaf,
Please, have a look at #46, I have clarified the CONTRIBUTING instructions.

Regarding your question:

> But I though that only OpenMM was not available with pip, right? What prevents you from changing tox.ini from

It is just a matter of time, testing and CI becomes much faster if installing dependencies via anaconda. This is also stated in the current version of the documentation.

I wait for your comment before closing and merging the PR.
