### 2020-05-23 joaomcteixeira/taurenmd#45 @joaomcteixeira
Thanks @Luthaf  for your feedback. I will address these questions directly in a new draft of the paper. I will let you know when committed. (ping also to @richardjgowers)
