### 2020-05-23 261ecbb
updates paper according to #45
