+++ joss/paper.md
@@ -23,15 +23,17 @@ bibliography: paper.bib
 
 # Summary
 
-Molecular dynamics (MD) simulations of biological molecules have evolved drastically since its application was first demonstrated four decades ago [@Karplus1977] and, nowadays, simulation of systems comprising millions of atoms is possible due to the latest advances in computation and data storage capacity -- and the scientific community's interest is growing [@Hospital2019]. Academic groups develop most of the MD methods and software for MD data handling and analysis. The MD analysis libraries developed solely for the latter scope nicely address the needs of manipulating raw data and calculating structural parameters [@mda1; @mda2; @mdt; @Romo2014; @Roe2013], each with its advantages and drawbacks inherent to their implementation strategies. This diversity enriches the field with a panoply of strategies that the community can utilize.
+Molecular dynamics (MD) simulations of biological molecules have evolved drastically since its application was first demonstrated four decades ago [@Karplus1977] and, nowadays, simulation of systems comprising millions of atoms is possible due to the latest advances in computation and data storage capacity -- and the scientific community's interest is growing [@Hospital2019]. Academic groups develop most of the MD methods and software for MD data handling and analysis. The MD analysis libraries developed solely for the latter scope nicely address the needs of manipulating raw data and calculating structural parameters, such as: `MDAnalysis` [@mda1; @mda2]; `MDTraj` [@mdt]; `LOOS` [@Romo2014]; and `PyTraj` [@PYTRAJ; @Roe2013], each with its advantages and drawbacks inherent to their implementation strategies. This diversity enriches the field with a panoply of strategies that the community can utilize.
 
-The MD analysis software libraries widely distributed and adopted by the community share two main characteristics: 1) they are written in pure Python [@CS-R9526], or provide a Python interface; and 2) they are *libraries*: highly versatile and powerful pieces of software that, however, require advanced scripting and understanding to be operated, even for their basic functionalities. While this is the correct approach to develop flexible computational libraries, it creates a barrier between these software packages and the *non-developer* researcher or high throughput practices, particularly for routine data handling. Therefore, the need has emerged to create a platform that efficiently combines the MD libraries available in the Python universe, taking the most out of each, and implements rapid interfaces for routine usage by both experts and non-experts in the field. In response to that, here is presented **taurenmd** (\autoref{fig:logo}), an easy-to-use and extensible ecosystem of command-line interfaces that facilitates complex operations in Molecular Dynamics data analysis by building on top of powerful Python-based MD analysis libraries.
+The MD analysis software libraries widely distributed and adopted by the community share two main characteristics: 1) they are written in pure Python [@CS-R9526], or provide a Python interface; and 2) they are *libraries*: highly versatile and powerful pieces of software that, however, require advanced scripting and understanding to be operated, even for their basic functionalities. While this is the correct approach to develop flexible computational libraries, it creates a barrier between these software packages and *non-developer* researchers and hinder high throughput practices, particularly for routine data handling. Therefore, the need has emerged to create a platform that can efficiently combine the MD libraries available in the Python universe, taking the most out of each, and implements rapid interfaces for routine usage by both experts and non-experts in the field. In response to that, here is presented **taurenmd** (\autoref{fig:logo}), an easy-to-use and extensible ecosystem of command-line interfaces that facilitates complex operations in Molecular Dynamics data analysis by building on top of powerful Python-based MD analysis libraries.
+
[…]
