+++ joss/paper.md
@@ -27,7 +27,7 @@ Molecular dynamics (MD) simulations of biological molecules have evolved drastic
 
 The MD analysis software libraries widely distributed and adopted by the community share two main characteristics: 1) they are written in pure Python [@CS-R9526], or provide a Python interface; and 2) they are *libraries*: highly versatile and powerful pieces of software that, however, require advanced scripting and understanding to be operated, even for their basic functionalities. While this is the correct approach to develop flexible computational libraries, it creates a barrier between these software packages and *non-developer* researchers and hinder high throughput practices, particularly for routine data handling. Therefore, the need has emerged to create a platform that can efficiently combine the MD libraries available in the Python universe, taking the most out of each, and implements rapid interfaces for routine usage by both experts and non-experts in the field. In response to that, here is presented **taurenmd** (\autoref{fig:logo}), an easy-to-use and extensible ecosystem of command-line interfaces that facilitates complex operations in Molecular Dynamics data analysis by building on top of powerful Python-based MD analysis libraries.
 
-The MD community has software packages that provide graphical interfaces to manipulate trajectories iteratively, such as VMD [@VMD], and the new integration between MDAnalysis and PyMol [@MDAPYMOL]. Also, there are powerful libraries to prepare molecular systems and run MD simulations, such as GROMACS [@GROMACS] and PLUMED [@PLUMED]. To date, *taurenmd* does not operates in either of these realms. If similar functionalities are developed in future versions, *taurenmd* will make use of the library interfaces provided by the projects referred before and will do so by creating command-line accessible pre-configured work-flows. Projects like GROMACS [@GROMACS] and LOOS [@LOOS] implement command-line ready interfaces to operate upon trajectories, differently to those libraries, *taurenmd* does not implement core functionalities and, instead, it leverages the features provided by the referred third-party libraries. Also, *taurenmd* is a Python written software and, therefore, focuses on the usage of MD libraries that are also written in Python or provide a Python interface.
+The MD community has software packages that provide graphical interfaces to manipulate trajectories iteratively, such as VMD [@VMD], and the new integration between MDAnalysis and PyMol [@MDAPYMOL]. Also, there are powerful libraries to prepare molecular systems and run MD simulations, such as GROMACS [@GROMACS] and PLUMED [@PLUMED]. To date, *taurenmd* does not operates in either of these realms. If similar functionalities are developed in future versions, *taurenmd* will make use of the library interfaces provided by the projects referred before and will do so by creating command-line accessible pre-configured work-flows. Projects like GROMACS [@GROMACS] and LOOS [@Romo2014] implement command-line ready interfaces to operate upon trajectories, differently to those libraries, *taurenmd* does not implement core functionalities and, instead, it leverages the features provided by the referred third-party libraries. Also, *taurenmd* is a Python written software and, therefore, focuses on the usage of MD libraries that are also written in Python or provide a Python interface.
 
 ![taurenmd logo.\label{fig:logo}](../docs/logo/taurenmd_logo_black_readme.png)
 
