### 2020-05-23 joaomcteixeira/taurenmd#45 @joaomcteixeira
> Whenever I see a package telling me to install it with pip, I assume it works fully on PyPi, with no dependencies on conda universe; and inversely. I am sure some future users of taurenmd will make the same distinction.

You are right. I will look for it in future versions when I merge the development branches currently open. Do you think that is an issue problematic enough to hinder publication?

I have addressed your other comments in a new draft that is now available -> https://github.com/openjournals/joss-reviews/issues/2175#issuecomment-633145301
