### 2020-05-23 #2175 @joaomcteixeira
@richardjgowers @Luthaf

The new article version addresses the issues referred by @Luthaf in https://github.com/joaomcteixeira/taurenmd/issues/45

I look forward to your comments, Thanks.
