### 2020-05-25 joaomcteixeira/taurenmd#45 @Luthaf
The changes to the paper look good to me.

> Do you think that is an issue problematic enough to hinder publication?

No, I think taurenmd is fine as is for publication 😃; this is mostly an improvement suggestion.
