### 2020-05-25 joaomcteixeira/taurenmd#41 @Luthaf
The changes in #46 looks better!
