### 2020-05-25 #2175 @Luthaf
With the latest changes to documentation done by @joaomcteixeira, I am happy to recommend this paper for publication in JOSS!
