### 2020-05-25 joaomcteixeira/taurenmd#41 @joaomcteixeira
Thanks for your input and help @Luthaf, I am merging to it master now. :wink:
Closed by #46
