### 2020-05-25 joaomcteixeira/taurenmd#45 @joaomcteixeira
:smile: Thanks for all your inputs and review process, it was very productive! Hope to see you around more often, cheers! :wink:
