### 2020-05-25 #2175 @joaomcteixeira
Thanks to the reviewers and the editors for all their help and insights. This review process was very fruitful, indeed! And I enjoy much this online and open process in JOSS, it really pushes scientific software development further.
