### 2020-05-26 35d3cee
improved traj reporting in MDAnalysis
