### 2020-05-31 #2175 @arfon
> @openjournals/joss-eics this is good to go!

Thanks @richardjgowers. In order to notify the EiC team (and to create final proofs) you need to as @whedon to `@whedon accept`. That said, it looks like this submission still needs an archive DOI for the software.
