### 2020-05-31 #2175 @joaomcteixeira
@arfon a quick question before doing anything, sorry for making you repeat, but I want to double-check beforehand.
* all the additions from this review are already part of the deployed `taurenmd`, currently version `v0.9.3`, which was done immediately after each PR for each new addition.
* new versions are automatically deployed in Zenodo (currently,  DOI: 10.5281/zenodo.3843326 ) -> https://zenodo.org/record/3843326#.XtQ3ORZq3ys
* latest git release: https://github.com/joaomcteixeira/taurenmd/releases

Do I still need to pack another Zenodo release?

Thank you
