### 2020-06-01 #2175 @kyleniemeyer
Hi @joaomcteixeira, just a few minor comments on the paper before I publish it:

- typo in the third paragraph: "To date, taurenmd does not operates in" -> "To date, taurenmd does not operate in"
- in the first paragraph of Implementation, "etc..." should just be "etc." (and not in italics
- same paragraph: "...; yet, simple, flat, easy to read, and extensible." should be something like "...; yet, it is simple, flat, easy to read, and extensible."
- please add the DOI to the Humphrey 1996 reference (10.1016/0263-7855(96)00018-5)

Once you've made these changes, please let us know here, and we'll get the article published!
