+++ joss/paper.bib
@@ -159,6 +159,7 @@ author={William Humphrey and Andrew Dalke and Klaus Schulten},
 title={{VMD} -- {V}isual {M}olecular {D}ynamics},
 journal={Journal of Molecular Graphics},
 year=1996,
+doi={10.1016/0263-7855(96)00018-5}
 volume=14,
 pages={33-38},
 note={},
