### 2020-06-02 #2175 @joaomcteixeira
Hi @kyleniemeyer ,
I have corrected your highlights in the article and the new version is already merged to `master`. It compiled properly using the JOSS preview tool. Do I need to create another Zenodo DOI with this new paper revision?
