### 2020-06-03 #2175 @joaomcteixeira
Dear @kyleniemeyer and @arfon,
I just noticed that in the article citation string (web, PDF bottom of pages, bibtex), as well as, in the article JOSS page, my name appears incomplete as 'Teixeira, J.M.' or 'João Teixeira', instead of my scientific full name: `Teixeira, J.M.C.` or `João M.C. Teixeira` (as in the main article authors section).

Is it possible to update these citation references to my full scientific name?

Thank you in advance.
