### 2020-06-03 #2175 @arfon
Thanks for flagging this @joaomcteixeira. This should now be fixed. Please note the PDF on the JOSS site will take some time to update as it is cached.
