### 2020-06-03 #2175 @joaomcteixeira
@arfon thank you very much, it looks okay now, also in the PDF file. I am happy to see that this resulted in an actual PR to improve the authors experience in JOSS.
Best wishes,
